small formatting fixes
