+++ src/components/HomePage/MonthAtAGlance.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Card, Alert, Button } from 'react-bootstrap';
-import BREAKTIME_BLUE from '../../utils';
-import { data, dataBar } from '../../utils';
+import { defaultColors } from '../../utils';
+import {data, dataBar} from './dummyData';
 
 import { IconContext } from 'react-icons';
 import { BsFillFileBarGraphFill } from 'react-icons/bs';
@@ -17,7 +17,7 @@ export default function MonthAtAGlance() {
   return (
     <div className='monthAtAGlance' style={{ 'display': 'flex', 'width': '50%' }}>
       <Card>
-        <Card.Header as='h5' style={{ 'backgroundColor': BREAKTIME_BLUE, 'color': 'white' }}>
+        <Card.Header as='h5' style={{ 'backgroundColor': defaultColors.BREAKTIME_BLUE, 'color': 'white' }}>
           <IconContext.Provider value={{ color: 'white' }}>
             <BsFillFileBarGraphFill />
           </IconContext.Provider>
